oblu (IMU) precisely guiding an arduino based robot
The robot moves in a pre-programmed path and transmits its current
coordinates to a phone for real-time tracking. An Android application
Xoblu is used for real-time tracking on phone. The robot's movement is
NOT remotely controlled, except start / stop commands. oblu (IMU) senses
motion and communicates movement information to Arduino. Based on the
programmed path and the motion information (sent by oblu), Arduino
controls wheels' movements.